+++ pycobranca/cnab/cnab400/__init__.py
@@ -10,6 +10,7 @@ from .citibank import RemessaCitibank400
 from .credisis import RemessaCredisis400
 from .inter import RemessaInter400
 from .itau import RemessaItau400
+from .safra import RemessaSafra400
 from .santander import RemessaSantander400
 from .sicoob import RemessaSicoob400
 from .unicred import RemessaUnicred400
@@ -28,4 +29,5 @@ __all__ = [
     "RemessaCredisis400",
     "RemessaInter400",
     "RemessaBancoC6_400",
+    "RemessaSafra400",
 ]